This changes was made at 12:50 by mohammed
Instead of parallel being in “info and about” i want it to be only
showing within the div class workk transperent img(which i called work
picture) i placed the background image that i want to use in the
parallel (as i scoll up denver city picture shows, as i scroll down the
code shows)

further more i didn’t like the fixed navigational bar and the padding
to 20% wasn’t working for me so i changed it to @media check the CSS to
see how i aligned the NavBAR to always be in the middle.

Thank you.