Last changes for 5/2/2014
Im done for the night, still problems with the navbar, i will come back
to it later. i haven’t had a chance to look at it. this time was mainly
organizing files and content inside html and css, giving it a clean
look thats easy to make changes to if ever needed. so far I’m happy
with the way things are turning out.